requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
mysql-connector-python==8.0.26
python-dotenv==0.19.0
openai==0.27.0
flask-login==0.5.0
Werkzeug==2.0.1
PyMySQL==1.0.2
python-dateutil==2.8.2